feat: updates to friendslistcomp and itemgroupcomp
